integrate admin features to monitoring page

+++ public_html/post.php
@@ -43,10 +43,9 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 
         $admin_presence = $database_helper->get("SELECT * FROM admins WHERE username = '$admin_username';");
 
-        // CHANGE TO > 0 once admins are implemented
-        // if (count($admin_presence) >= 0) {
-        $database_helper->set("INSERT INTO posts (`id`, `username`, `post_title`, `post_content`, `timestamp`) VALUES (DEFAULT, '$admin_username', '$post_title', '$post_content', DEFAULT);");
-        // }
+        if (count($admin_presence) >= 0) {
+            $database_helper->set("INSERT INTO posts (`id`, `username`, `post_title`, `post_content`, `timestamp`) VALUES (DEFAULT, '$admin_username', '$post_title', '$post_content', DEFAULT);");
+        }
     } else if (isset($_POST['system-notification-delete'])) {
         /// --------------------------------
         /// DETECT A POST DELETION REQUEST
@@ -56,10 +55,9 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
         $post_id = filter_var($_POST['system-notification-delete-id'], FILTER_SANITIZE_STRING);
         $admin_presence = $database_helper->get("SELECT * FROM admins WHERE username = '$admin_username';");
 
-        // CHANGE TO > 0 once admins are implemented
-        // if (count($admin_presence) >= 0) {
-        $database_helper->set("DELETE FROM posts WHERE id=$post_id;");
-        // }
+        if (count($admin_presence) >= 0) {
+            $database_helper->set("DELETE FROM posts WHERE id=$post_id;");
+        }
     }
 
     header('Location: status.php', true, 301);
